alpha-starter-workspace/alpha-api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
pydantic[email]
pydantic-settings>=2.2
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary==2.9.9
alembic
httpx
redis
bcrypt
PyJWT
email-validator
openai>=1.52.0
pywebpush
cryptography
